Move gdb process into src/gdb.rs
